requirements.txt
openai>=1.0
python-dotenv>=1.0
pymupdf>=1.23
numpy>=1.24
python-docx>=1.1
beautifulsoup4>=4.12
pyyaml>=6.0
sentence-transformers>=2.2
umap-learn>=0.5
matplotlib>=3.7
scipy>=1.11
